nested comment actions inside single button to save space
